Rebrand to gver.org, closes #63
There are just too many other gitVersion plugins out there and the
domains are already taken. In order to separate this a bit from the
rest, the plugin will be developed under gver.org. This is the last
release with the pluginId org.dmfs.gitversion. The next release will
have the ID `org.gver`.